Reorganize record tests.
