+++ db/tests/records/operations/test_select.py
@@ -1,0 +1,20 @@
+from db.records.operations.select import get_records
+
+
+def test_get_records_gets_all_records(roster_table_obj):
+    roster, engine = roster_table_obj
+    record_list = get_records(roster, engine)
+    assert len(record_list) == 1000
+
+
+def test_get_records_gets_limited_records(roster_table_obj):
+    roster, engine = roster_table_obj
+    record_list = get_records(roster, engine, limit=10)
+    assert len(record_list) == 10
+
+
+def test_get_records_gets_limited_offset_records(roster_table_obj):
+    roster, engine = roster_table_obj
+    base_records = get_records(roster, engine, limit=10)
+    offset_records = get_records(roster, engine, limit=10, offset=5)
+    assert len(offset_records) == 10 and offset_records[0] == base_records[5]
